LPD-6742 Improving virtual instance migrator extraction module unit tests by reusing more code
