+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-extractor/source-formatter-suppressions.xml
@@ -4,5 +4,6 @@
 	<checkstyle>
 		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/extractor/util/DatabaseUtil\.java" />
 		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/extractor/DBPartitionVirtualInstanceMigratorExtractorTest\.java" />
+		<suppress checks="CompanyIterationCheck" files="src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/extractor/util/DatabaseMockupUtil\.java" />
 	</checkstyle>
 </suppressions>
