+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-extractor/src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/extractor/DBPartitionVirtualInstanceMigratorExtractorTest.java
@@ -10,21 +10,17 @@ import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.version.Version;
 import com.liferay.portal.tools.db.partition.virtual.instance.migrator.common.Company;
 import com.liferay.portal.tools.db.partition.virtual.instance.migrator.common.Release;
+import com.liferay.portal.tools.db.partition.virtual.instance.migrator.extractor.util.DatabaseMockupUtil;
 
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 
-import java.sql.Connection;
-import java.sql.DatabaseMetaData;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 
-import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import org.junit.Assert;
@@ -32,184 +28,68 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import org.mockito.MockedStatic;
-import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
-
 /**
  * @author Luis Ortiz
  */
-public class DBPartitionVirtualInstanceMigratorExtractorTest {
+public class DBPartitionVirtualInstanceMigratorExtractorTest
+	extends DatabaseMockupUtil {
 
 	@Test
 	public void testMultipleCompanyDefaultDatabase()
 		throws IOException, SQLException {
 
-		List<Long> companyInfoIds = new ArrayList<>();
-
-		companyInfoIds.add(RandomTestUtil.randomLong());
-		companyInfoIds.add(RandomTestUtil.randomLong());
-
-		boolean defaultPartition = true;
-
-		List<Long> companyIds = new ArrayList<>();
-
-		companyIds.add(RandomTestUtil.randomLong());
-		companyIds.add(RandomTestUtil.randomLong());
-
-		Release module1Release = new Release(
-			Version.parseVersion("14.2.4"), "module1", 0, true);
-		Release module2Release = new Release(
-			Version.parseVersion("2.0.1"), "module2", 1, false);
-
-		List<Release> releases = new ArrayList<>();
-
-		releases.add(module1Release);
-		releases.add(module2Release);
-
-		List<Company> companies = new ArrayList<>();
-
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-
 		_executeTest(
-			companyInfoIds, defaultPartition, companyIds, releases, companies);
+			_generateCompanies(),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			true, _generateReleases());
 	}
 
 	@Test
 	public void testMultipleCompanyNondefaultDatabase()
 		throws IOException, SQLException {
 
-		List<Long> companyInfoIds = new ArrayList<>();
-
-		companyInfoIds.add(RandomTestUtil.randomLong());
-		companyInfoIds.add(RandomTestUtil.randomLong());
-
-		boolean defaultPartition = false;
-
-		List<Long> companyIds = new ArrayList<>();
-
-		companyIds.add(RandomTestUtil.randomLong());
-		companyIds.add(RandomTestUtil.randomLong());
-
-		Release module1Release = new Release(
-			Version.parseVersion("14.2.4"), "module1", 0, true);
-		Release module2Release = new Release(
-			Version.parseVersion("2.0.1"), "module2", 1, false);
-
-		List<Release> releases = new ArrayList<>();
-
-		releases.add(module1Release);
-		releases.add(module2Release);
-
-		List<Company> companies = new ArrayList<>();
-
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-
 		_executeTest(
-			companyInfoIds, defaultPartition, companyIds, releases, companies);
+			_generateCompanies(),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			false, _generateReleases());
 	}
 
 	@Test
 	public void testSingleCompanyDefaultDatabase()
 		throws IOException, SQLException {
 
-		List<Long> companyInfoIds = new ArrayList<>();
-
-		companyInfoIds.add(RandomTestUtil.randomLong());
-
-		boolean defaultPartition = true;
-
-		List<Long> companyIds = new ArrayList<>();
-
-		companyIds.add(RandomTestUtil.randomLong());
-		companyIds.add(RandomTestUtil.randomLong());
-
-		Release module1Release = new Release(
-			Version.parseVersion("14.2.4"), "module1", 0, true);
-		Release module2Release = new Release(
-			Version.parseVersion("2.0.1"), "module2", 1, false);
-
-		List<Release> releases = new ArrayList<>();
-
-		releases.add(module1Release);
-		releases.add(module2Release);
-
-		List<Company> companies = new ArrayList<>();
-
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-
 		_executeTest(
-			companyInfoIds, defaultPartition, companyIds, releases, companies);
+			_generateCompanies(),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			Collections.singletonList(RandomTestUtil.randomLong()), true,
+			_generateReleases());
 	}
 
 	@Test
 	public void testSingleCompanyNondefaultDatabase()
 		throws IOException, SQLException {
 
-		List<Long> companyInfoIds = new ArrayList<>();
-
-		companyInfoIds.add(RandomTestUtil.randomLong());
-
-		boolean defaultPartition = false;
-
-		List<Long> companyIds = new ArrayList<>();
-
-		companyIds.add(RandomTestUtil.randomLong());
-		companyIds.add(RandomTestUtil.randomLong());
-
-		Release module1Release = new Release(
-			Version.parseVersion("14.2.4"), "module1", 0, true);
-		Release module2Release = new Release(
-			Version.parseVersion("2.0.1"), "module2", 1, false);
-
-		List<Release> releases = new ArrayList<>();
-
-		releases.add(module1Release);
-		releases.add(module2Release);
-
-		List<Company> companies = new ArrayList<>();
-
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-		companies.add(
-			new Company(
-				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
-
 		_executeTest(
-			companyInfoIds, defaultPartition, companyIds, releases, companies);
+			_generateCompanies(),
+			Arrays.asList(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomLong()),
+			Collections.singletonList(RandomTestUtil.randomLong()), false,
+			_generateReleases());
 	}
 
 	@Rule
 	public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
 	private void _assertFileContent(
-		String content, List<Long> companyInfoIds, boolean defaultPartition,
-		List<Release> releases, List<Company> companies) {
+		List<Company> companies, List<Long> companyInfoIds, String content,
+		boolean defaultPartition, List<Release> releases) {
 
 		content = content.replaceAll("\n", "");
 		content = content.replaceAll("\r", "");
@@ -241,13 +121,14 @@ public class DBPartitionVirtualInstanceMigratorExtractorTest {
 	}
 
 	private void _executeTest(
+			List<Company> companies, List<Long> companyIds,
 			List<Long> companyInfoIds, boolean defaultPartition,
-			List<Long> companyIds, List<Release> releases,
-			List<Company> companies)
+			List<Release> releases)
 		throws IOException, SQLException {
 
 		_mockDatabase(
-			companyInfoIds, defaultPartition, companyIds, releases, companies);
+			companies, companyIds, companyInfoIds, defaultPartition, releases,
+			Arrays.asList("Table1", "Company", "Table2", "Object_x_25000"));
 
 		File outputDirectory = temporaryFolder.newFolder("tempExports");
 
@@ -264,7 +145,23 @@ public class DBPartitionVirtualInstanceMigratorExtractorTest {
 		String content = _getFileContent(files[0]);
 
 		_assertFileContent(
-			content, companyInfoIds, defaultPartition, releases, companies);
+			companies, companyInfoIds, content, defaultPartition, releases);
+	}
+
+	private List<Company> _generateCompanies() {
+		return Arrays.asList(
+			new Company(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
+				RandomTestUtil.randomString(), RandomTestUtil.randomString()),
+			new Company(
+				RandomTestUtil.randomLong(), RandomTestUtil.randomString(),
+				RandomTestUtil.randomString(), RandomTestUtil.randomString()));
+	}
+
+	private List<Release> _generateReleases() {
+		return Arrays.asList(
+			new Release(Version.parseVersion("14.2.4"), "module1", 0, true),
+			new Release(Version.parseVersion("2.0.1"), "module2", 1, false));
 	}
 
 	private String _getCompaniesOutput(List<Company> companies) {
@@ -361,602 +258,19 @@ public class DBPartitionVirtualInstanceMigratorExtractorTest {
 		return sb.toString();
 	}
 
-	private void _mockCompanies(List<Company> companies) throws SQLException {
-		PreparedStatement preparedStatement = Mockito.mock(
-			PreparedStatement.class);
-		ResultSet resultSet = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_connection.prepareStatement(
-				"select Company.companyId, webId, name, hostname from " +
-					"Company left join VirtualHost on Company.companyId = " +
-						"VirtualHost.companyId")
-		).thenReturn(
-			preparedStatement
-		);
-
-		Mockito.when(
-			preparedStatement.executeQuery()
-		).thenReturn(
-			resultSet
-		);
-
-		final List<Integer> nextCounter = new ArrayList<>();
-
-		nextCounter.add(0);
-
-		Mockito.when(
-			resultSet.next()
-		).thenAnswer(
-			new Answer<Boolean>() {
-
-				@Override
-				public Boolean answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = nextCounter.get(0);
-
-					if (counter >= companies.size()) {
-						return false;
-					}
-
-					nextCounter.set(0, ++counter);
-
-					return true;
-				}
-
-			}
-		);
-
-		final List<Integer> companyIdCounter = new ArrayList<>();
-
-		companyIdCounter.add(0);
-
-		Mockito.when(
-			resultSet.getLong(1)
-		).thenAnswer(
-			new Answer<Long>() {
-
-				@Override
-				public Long answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = companyIdCounter.get(0);
-
-					if (counter >= companies.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					companyIdCounter.set(0, ++counter);
-
-					return companies.get(
-						counter - 1
-					).getCompanyId();
-				}
-
-			}
-		);
-
-		final List<Integer> webIdCounter = new ArrayList<>();
-
-		webIdCounter.add(0);
-
-		Mockito.when(
-			resultSet.getString(2)
-		).thenAnswer(
-			new Answer<String>() {
-
-				@Override
-				public String answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = webIdCounter.get(0);
-
-					if (counter >= companies.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					webIdCounter.set(0, ++counter);
-
-					return companies.get(
-						counter - 1
-					).getWebId();
-				}
-
-			}
-		);
-
-		final List<Integer> companyNameCounter = new ArrayList<>();
-
-		companyNameCounter.add(0);
-
-		Mockito.when(
-			resultSet.getString(3)
-		).thenAnswer(
-			new Answer<String>() {
-
-				@Override
-				public String answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = companyNameCounter.get(0);
-
-					if (counter >= companies.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					companyNameCounter.set(0, ++counter);
-
-					return companies.get(
-						counter - 1
-					).getCompanyName();
-				}
-
-			}
-		);
-
-		final List<Integer> virtualHostCounter = new ArrayList<>();
-
-		virtualHostCounter.add(0);
-
-		Mockito.when(
-			resultSet.getString(4)
-		).thenAnswer(
-			new Answer<String>() {
-
-				@Override
-				public String answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = virtualHostCounter.get(0);
-
-					if (counter >= companies.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					virtualHostCounter.set(0, ++counter);
-
-					return companies.get(
-						counter - 1
-					).getVirtualHostName();
-				}
-
-			}
-		);
-	}
-
 	private void _mockDatabase(
+			List<Company> companies, List<Long> companyIds,
 			List<Long> companyInfoIds, boolean defaultPartition,
-			List<Long> companyIds, List<Release> releases,
-			List<Company> companies)
+			List<Release> releases, List<String> tableNames)
 		throws SQLException {
 
-		_driverManagerMockedStatic.when(
-			() -> DriverManager.getConnection(_URL, _USER, _PASSWORD)
-		).thenReturn(
-			_connection
-		);
-
-		Mockito.when(
-			_connection.getMetaData()
-		).thenReturn(
-			_databaseMetaData
-		);
-
-		Mockito.when(
-			_databaseMetaData.getURL()
-		).thenReturn(
-			_URL
-		);
-
-		_mockGetCompanyId(companyInfoIds);
-		_mockDefaultPartition(defaultPartition);
-		_mockGetCompanyIds(companyIds);
-		_mockTables();
-		_mockReleases(releases);
-		_mockCompanies(companies);
-	}
-
-	private void _mockDefaultPartition(boolean defaultPartition)
-		throws SQLException {
-
-		ResultSet resultSet = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_databaseMetaData.getTables(
-				Mockito.nullable(String.class), Mockito.nullable(String.class),
-				Mockito.eq("Company"), Mockito.any(String[].class))
-		).thenReturn(
-			resultSet
-		);
-
-		Mockito.when(
-			resultSet.next()
-		).thenReturn(
-			defaultPartition
-		);
-	}
-
-	private void _mockGetCompanyId(List<Long> companyIds) throws SQLException {
-		PreparedStatement preparedStatement = Mockito.mock(
-			PreparedStatement.class);
-		ResultSet resultSet = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_connection.prepareStatement("select companyId from CompanyInfo")
-		).thenReturn(
-			preparedStatement
-		);
-
-		Mockito.when(
-			preparedStatement.executeQuery()
-		).thenReturn(
-			resultSet
-		);
-
-		final List<Integer> nextCounter = new ArrayList<>();
-
-		nextCounter.add(0);
-
-		Mockito.when(
-			resultSet.next()
-		).thenAnswer(
-			new Answer<Boolean>() {
-
-				@Override
-				public Boolean answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = nextCounter.get(0);
-
-					if (counter >= companyIds.size()) {
-						return false;
-					}
-
-					nextCounter.set(0, ++counter);
-
-					return true;
-				}
-
-			}
-		);
-
-		final List<Integer> getLongCounter = new ArrayList<>();
-
-		getLongCounter.add(0);
-
-		Mockito.when(
-			resultSet.getLong(1)
-		).thenAnswer(
-			new Answer<Long>() {
-
-				@Override
-				public Long answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = getLongCounter.get(0);
-
-					if (counter >= companyIds.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					getLongCounter.set(0, ++counter);
-
-					return companyIds.get(counter - 1);
-				}
-
-			}
-		);
-	}
-
-	private void _mockGetCompanyIds(List<Long> companyIds) throws SQLException {
-		PreparedStatement preparedStatement = Mockito.mock(
-			PreparedStatement.class);
-		ResultSet resultSet = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_connection.prepareStatement("select companyId from Company")
-		).thenReturn(
-			preparedStatement
-		);
-
-		Mockito.when(
-			preparedStatement.executeQuery()
-		).thenReturn(
-			resultSet
-		);
-
-		final List<Integer> nextCounter = new ArrayList<>();
-
-		nextCounter.add(0);
-
-		Mockito.when(
-			resultSet.next()
-		).thenAnswer(
-			new Answer<Boolean>() {
-
-				@Override
-				public Boolean answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = nextCounter.get(0);
-
-					if (counter >= companyIds.size()) {
-						return false;
-					}
-
-					nextCounter.set(0, ++counter);
-
-					return true;
-				}
-
-			}
-		);
-
-		final List<Integer> getLongCounter = new ArrayList<>();
-
-		getLongCounter.add(0);
-
-		Mockito.when(
-			resultSet.getLong("companyId")
-		).thenAnswer(
-			new Answer<Long>() {
-
-				@Override
-				public Long answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = getLongCounter.get(0);
-
-					if (counter >= companyIds.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					getLongCounter.set(0, ++counter);
-
-					return companyIds.get(counter - 1);
-				}
-
-			}
-		);
-	}
-
-	private void _mockReleases(List<Release> releases) throws SQLException {
-		PreparedStatement preparedStatement = Mockito.mock(
-			PreparedStatement.class);
-
-		Mockito.when(
-			_connection.prepareStatement(
-				"select servletContextName, schemaVersion, state_, verified " +
-					"from Release_")
-		).thenReturn(
-			preparedStatement
-		);
-
-		ResultSet resultSet = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			preparedStatement.executeQuery()
-		).thenReturn(
-			resultSet
-		);
-
-		final List<Integer> nextCounter = new ArrayList<>();
-
-		nextCounter.add(0);
-
-		Mockito.when(
-			resultSet.next()
-		).thenAnswer(
-			new Answer<Boolean>() {
-
-				@Override
-				public Boolean answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = nextCounter.get(0);
-
-					if (counter >= releases.size()) {
-						return false;
-					}
-
-					nextCounter.set(0, ++counter);
-
-					return true;
-				}
-
-			}
-		);
-
-		final List<Integer> servletContextNameCounter = new ArrayList<>();
-
-		servletContextNameCounter.add(0);
-
-		Mockito.when(
-			resultSet.getString(1)
-		).thenAnswer(
-			new Answer<String>() {
-
-				@Override
-				public String answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = servletContextNameCounter.get(0);
-
-					if (counter >= releases.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					servletContextNameCounter.set(0, ++counter);
-
-					return releases.get(
-						counter - 1
-					).getServletContextName();
-				}
-
-			}
-		);
-
-		final List<Integer> schemaVersionCounter = new ArrayList<>();
-
-		schemaVersionCounter.add(0);
-
-		Mockito.when(
-			resultSet.getString(2)
-		).thenAnswer(
-			new Answer<String>() {
-
-				@Override
-				public String answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = schemaVersionCounter.get(0);
-
-					if (counter >= releases.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					schemaVersionCounter.set(0, ++counter);
-
-					Version version = releases.get(
-						counter - 1
-					).getSchemaVersion();
-
-					return version.toString();
-				}
-
-			}
-		);
-
-		final List<Integer> stateCounter = new ArrayList<>();
-
-		stateCounter.add(0);
-
-		Mockito.when(
-			resultSet.getInt(3)
-		).thenAnswer(
-			new Answer<Integer>() {
-
-				@Override
-				public Integer answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = stateCounter.get(0);
-
-					if (counter >= releases.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					stateCounter.set(0, ++counter);
-
-					return releases.get(
-						counter - 1
-					).getState();
-				}
-
-			}
-		);
-
-		final List<Integer> verifiedCounter = new ArrayList<>();
-
-		verifiedCounter.add(0);
-
-		Mockito.when(
-			resultSet.getBoolean(4)
-		).thenAnswer(
-			new Answer<Boolean>() {
-
-				@Override
-				public Boolean answer(InvocationOnMock invocationOnMock)
-					throws Throwable {
-
-					int counter = verifiedCounter.get(0);
-
-					if (counter >= releases.size()) {
-						throw new IndexOutOfBoundsException();
-					}
-
-					verifiedCounter.set(0, ++counter);
-
-					return releases.get(
-						counter - 1
-					).getVerified();
-				}
-
-			}
-		);
-	}
-
-	private void _mockTables() throws SQLException {
-		ResultSet resultSet1 = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_databaseMetaData.getColumns(
-				Mockito.nullable(String.class), Mockito.nullable(String.class),
-				Mockito.any(), Mockito.nullable(String.class))
-		).thenReturn(
-			resultSet1
-		);
-
-		Mockito.when(
-			resultSet1.next()
-		).thenReturn(
-			true
-		);
-
-		/*ResultSet resultSet2 = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_databaseMetaData.getColumns(
-				Mockito.nullable(String.class), Mockito.nullable(String.class),
-				Mockito.eq("Company"), Mockito.nullable(String.class))
-		).thenReturn(
-			resultSet2
-		);
-
-		Mockito.when(
-			resultSet2.next()
-		).thenReturn(
-			true
-		);*/
-
-		ResultSet resultSet3 = Mockito.mock(ResultSet.class);
-
-		Mockito.when(
-			_databaseMetaData.getTables(
-				Mockito.nullable(String.class), Mockito.nullable(String.class),
-				Mockito.isNull(), Mockito.any(String[].class))
-		).thenReturn(
-			resultSet3
-		);
-
-		// Mock _resultSet
-
-		Mockito.when(
-			resultSet3.getString("TABLE_NAME")
-		).thenReturn(
-			"Table1"
-		).thenReturn(
-			"Company"
-		).thenReturn(
-			"Table2"
-		).thenReturn(
-			"Object_x_25000"
-		);
-
-		Mockito.when(
-			resultSet3.next()
-		).thenReturn(
-			true
-		).thenReturn(
-			true
-		).thenReturn(
-			true
-		).thenReturn(
-			true
-		).thenReturn(
-			false
-		);
+		mockCompanies(companies);
+		mockDatabaseConnection(_PASSWORD, _URL, _USER);
+		mockDefaultPartition(defaultPartition);
+		mockGetCompanyIds(companyIds);
+		mockGetCompanyInfoIds(companyInfoIds);
+		mockReleases(releases);
+		mockTables(tableNames);
 	}
 
 	private static final String _PASSWORD = RandomTestUtil.randomString();
@@ -964,12 +278,5 @@ public class DBPartitionVirtualInstanceMigratorExtractorTest {
 	private static final String _URL = RandomTestUtil.randomString();
 
 	private static final String _USER = RandomTestUtil.randomString();
-
-	private static final MockedStatic<DriverManager>
-		_driverManagerMockedStatic = Mockito.mockStatic(DriverManager.class);
-
-	private final Connection _connection = Mockito.mock(Connection.class);
-	private final DatabaseMetaData _databaseMetaData = Mockito.mock(
-		DatabaseMetaData.class);
 
 }
